requirements.txt
fastapi[all]
psycopg2
sqlalchemy
python-jose[cryptography]